+++ tests/e2e/cases/event_emitter_max_listeners/in.ts
@@ -1,8 +1,13 @@
 import { EventEmitter, getMaxListeners, setMaxListeners } from "events";
 import * as nodeEvents from "node:events";
 
-const emitter = new EventEmitter();
-const other = new nodeEvents.EventEmitter();
+let ignoredSeen = "";
+function mark(label: string): string {
+    ignoredSeen += label;
+    return label;
+}
+const emitter = new EventEmitter(mark("c"));
+const other = new nodeEvents.EventEmitter(mark("n"));
 
 console.log("default:", emitter.getMaxListeners());
 console.log("chain:", emitter.setMaxListeners(2).getMaxListeners());
@@ -10,3 +15,4 @@ setMaxListeners(4, emitter);
 console.log("static:", getMaxListeners(emitter));
 nodeEvents.setMaxListeners(6, other);
 console.log("namespace:", nodeEvents.getMaxListeners(other));
+console.log("ignored:", emitter.getMaxListeners(mark("g")), other.getMaxListeners(mark("h")), ignoredSeen);